User Management v.2.0 (previously Pin Management)
Pin Management is now User Management. Employee now have their own class to store data for each.
Timeclock data is now stored directly as a part of the Employee class. Variables with "kronos" in the name refer to timeclock-related functionality.
_CreateUser()_, _EditUser()_, _DeleteUser()_ have been stripped down and simplified (these were previously _CreatePin()_, _EditPin()_, _DeletePin()_). The previous menu is now only a debug menu for testing without a UI.
Authenticate() function is being simplified but currently has no data to iterate over so the loop doesn't work. It should return the access value for the given pin's role (based on the Configurable Constants.
Possible planned implementation of Validate() functions that will run the checks when called and return a value instead of the code being run inside of the Create/Edit/Delete/Authenticate functions.

*Problems:*
Figuring out a way to store all of the necessary data in a .csv and then load that without lots of conversion. Load() should just be able to scan every line of the database for the information it needs and then use those to call the Employee constructor on each line. Right now the data is not in the correct format to work. Also figuring out how to convert ctime values to readable strings and back is a problem. Not sure whether to store ctime values or readable time values. There is some non-functional remnants from the last 4 attempts to implement read/write between the .csv file and the Employee class.
Destructor doesn't work yet because it has to use pointers.