Added very basic Cook ClockIn,ClockOut, and GetCheck functionality
Clock In and Clock Out for the cook menu writes the times on 2 separate files per user. The function Get Check then subtracts the 2 times to determine the amount of time worked. It then takes the time worked and figures out the check amount @$7.25 an hour.

Not very intuitive but its something.

It does not have error checking for:

-If a brand new user Clocks Out before Clocking In. If this brand new user clicks Get Check then they get a negative amount.

-If user Clocks Out but never Clocked In during that instance, when they click Get Check, it will calculate the time they worked based on their Previous Clock In time with their current Clock Out time, resulting in a bigger check.

-User can click Get Check multiple times.

To reset the clock in/clock out, just delete all the .txt files